Añadido log niveles DEBUG e INFO y corregidos Warnings y Errores
