+++ inejsonstat.log
@@ -1,0 +1,40 @@
+27/04/2022 10:15:54 PM; INFO; --------------------------------------------------------------------------------
+27/04/2022 10:15:54 PM; INFO; Starting program
+27/04/2022 10:15:54 PM; INFO; Executing module [build_url]
+27/04/2022 10:15:54 PM; DEBUG; Config file read successful
+27/04/2022 10:15:54 PM; INFO; Executing module [check_nult]
+27/04/2022 10:15:54 PM; DEBUG; No nult parameter
+27/04/2022 10:15:54 PM; INFO; Executing module [check_date]
+27/04/2022 10:15:54 PM; INFO; Date parameter valid, format: YYYYMMDD
+27/04/2022 10:15:54 PM; INFO; Executing module [check_input]
+27/04/2022 10:15:54 PM; INFO; URL parameterized working
+27/04/2022 10:15:54 PM; INFO; The URL is: https://servicios.ine.es/wstempus/jsstat/ES/DATASET/2065?date=20211001
+27/04/2022 10:15:54 PM; INFO; URL working, starting dataset generation
+27/04/2022 10:15:54 PM; INFO; Executing module [generate_object]
+27/04/2022 10:15:54 PM; INFO; Executing module [get_number_dimensions]
+27/04/2022 10:15:54 PM; DEBUG; Module [get_number_dimensions], limit position: 2, list index out of range
+27/04/2022 10:15:54 PM; INFO; Executing module [generate_dimensions]
+27/04/2022 10:15:54 PM; INFO; Executing module [generate_category]
+27/04/2022 10:15:54 PM; INFO; Executing module [calculate_category_size]
+27/04/2022 10:15:54 PM; DEBUG; Module [calculate_category_size], limit position: 63, dimension 'comunidadesautonomasyprovincias': unknown index or label '63'
+27/04/2022 10:15:54 PM; INFO; Size of category: 63
+27/04/2022 10:15:54 PM; INFO; Executing module [generate_index]
+27/04/2022 10:15:54 PM; INFO; Executing module [check_index]
+27/04/2022 10:15:54 PM; INFO; Executing module [check_label]
+27/04/2022 10:15:54 PM; INFO; Index: {0: '5_00', 1: '1_01', 2: '4_04', 3: '4_11', 4: '4_14', 5: '4_18', 6: '4_21', 7: '4_23', 8: '4_29', 9: '4_41', 10: '1_02', 11: '4_22', 12: '4_44', 13: '4_50', 14: '1_03', 15: '1_04~15', 16: '1_05', 17: '4_35', 18: '4_38', 19: '1_06', 20: '1_07', 21: '4_05~21', 22: '4_09', 23: '4_24', 24: '4_34', 25: '4_37', 26: '4_40', 27: '4_42', 28: '4_47', 29: '4_49', 30: '1_08', 31: '4_02~31', 32: '4_13', 33: '4_16', 34: '4_19', 35: '4_45', 36: '1_09~36', 37: '4_08~37', 38: '4_17', 39: '4_25', 40: '4_43', 41: '1_10', 42: '4_03~42', 43: '4_12', 44: '4_46', 45: '1_11~45', 46: '4_06~46', 47: '4_10~47', 48: '1_12~48', 49: '4_15', 50: '4_27', 51: '4_32', 52: '4_36', 53: '1_13~53', 54: '1_14~54', 55: '1_15~55', 56: '1_16~56', 57: '4_01~57', 58: '4_48', 59: '4_20', 60: '1_17~60', 61: '1_18~61', 62: '1_19~62'}
+27/04/2022 10:15:54 PM; INFO; Label: {'5_00': 'Total Nacional', '1_01': 'Andalucía', '4_04': 'Almería', '4_11': 'Cádiz', '4_14': 'Córdoba', '4_18': 'Granada', '4_21': 'Huelva', '4_23': 'Jaén', '4_29': 'Málaga', '4_41': 'Sevilla', '1_02': 'Aragón', '4_22': 'Huesca', '4_44': 'Teruel', '4_50': 'Zaragoza', '1_03': 'Asturias, Principado de', '1_04~15': 'Balears, Illes', '1_05': 'Canarias', '4_35': 'Palmas, Las', '4_38': 'Santa Cruz de Tenerife', '1_06': 'Cantabria', '1_07': 'Castilla y León', '4_05~21': 'Ávila', '4_09': 'Burgos', '4_24': 'León', '4_34': 'Palencia', '4_37': 'Salamanca', '4_40': 'Segovia', '4_42': 'Soria', '4_47': 'Valladolid', '4_49': 'Zamora', '1_08': 'Castilla - La Mancha', '4_02~31': 'Albacete', '4_13': 'Ciudad Real', '4_16': 'Cuenca', '4_19': 'Guadalajara', '4_45': 'Toledo', '1_09~36': 'Cataluña', '4_08~37': 'Barcelona', '4_17': 'Girona', '4_25': 'Lleida', '4_43': 'Tarragona', '1_10': 'Comunitat Valenciana', '4_03~42': 'Alicante/Alacant', '4_12': 'Castellón/Castelló', '4_46': 'Valencia/València', '1_11~45': 'Extremadura', '4_06~46': 'Badajoz', '4_10~47': 'Cáceres', '1_12~48': 'Galicia', '4_15': 'Coruña, A', '4_27': 'Lugo', '4_32': 'Ourense', '4_36': 'Pontevedra', '1_13~53': 'Madrid, Comunidad de', '1_14~54': 'Murcia, Región de', '1_15~55': 'Navarra, Comunidad Foral de', '1_16~56': 'País Vasco', '4_01~57': 'Araba/Álava', '4_48': 'Bizkaia', '4_20': 'Gipuzkoa', '1_17~60': 'Rioja, La', '1_18~61': 'Ceuta', '1_19~62': 'Melilla'}
+27/04/2022 10:15:54 PM; INFO; Executing module [generate_category]
+27/04/2022 10:15:54 PM; INFO; Executing module [calculate_category_size]
+27/04/2022 10:15:54 PM; DEBUG; Module [calculate_category_size], limit position: 1, dimension 'Per': unknown index or label '1'
+27/04/2022 10:15:54 PM; INFO; Size of category: 1
+27/04/2022 10:15:54 PM; INFO; Executing module [generate_index]
+27/04/2022 10:15:54 PM; INFO; Executing module [check_index]
+27/04/2022 10:15:54 PM; INFO; Executing module [check_label]
+27/04/2022 10:15:54 PM; INFO; Index: {0: '2021M10'}
+27/04/2022 10:15:54 PM; INFO; Label: {'2021M10': '2021M10'}
+27/04/2022 10:15:54 PM; INFO; Executing module [normalize_string]
+27/04/2022 10:15:54 PM; INFO; Executing module [normalize_string]
+27/04/2022 10:15:54 PM; INFO; Executing module [generate_value]
+27/04/2022 10:15:54 PM; DEBUG; Module [generate_value], limit position: 63, list index out of range
+27/04/2022 10:15:54 PM; INFO; Executing module [generate_status]
+27/04/2022 10:15:54 PM; DEBUG; Module [generate_status], limit position: 63, list index out of range
